+++ src/condor_includes/proc.h
@@ -93,8 +93,8 @@ END_C_DECLS
 // Put C++ definitions here
 #if defined(__cplusplus)
 bool operator==( const PROC_ID a, const PROC_ID b);
-unsigned int hashFuncPROC_ID( const PROC_ID & );
-unsigned int hashFunction(const PROC_ID &);
+size_t hashFuncPROC_ID( const PROC_ID & );
+size_t hashFunction(const PROC_ID &);
 void procids_to_mystring(ExtArray<PROC_ID> *procids, MyString &str);
 ExtArray<PROC_ID>* mystring_to_procids(MyString &str);
 
@@ -142,11 +142,11 @@ typedef struct JOB_ID_KEY {
 	operator std::string() const;
 	void sprint(MyString &s) const;
 	bool set(const char * job_id_str) { return StrIsProcId(job_id_str, this->cluster, this->proc, NULL); }
-	static unsigned int hash(const JOB_ID_KEY &);
+	static size_t hash(const JOB_ID_KEY &);
 } JOB_ID_KEY;
 
 inline bool operator==( const JOB_ID_KEY a, const JOB_ID_KEY b) { return a.cluster == b.cluster && a.proc == b.proc; }
-unsigned int hashFunction(const JOB_ID_KEY &);
+size_t hashFunction(const JOB_ID_KEY &);
 
 
 #endif
