+++ src/condor_utils/HashTable.cpp
@@ -22,7 +22,7 @@
 #include "HashTable.h"
 #include "proc.h"
 
-unsigned int
+size_t
 hashFuncInt( const int& n )
 {
 	if( n < 0 ) {
@@ -31,7 +31,7 @@ hashFuncInt( const int& n )
 	return n;
 }
 
-unsigned int
+size_t
 hashFuncLong( const long& n )
 {
 	if( n < 0 ) {
@@ -40,16 +40,16 @@ hashFuncLong( const long& n )
 	return n;
 }
 
-unsigned int
+size_t
 hashFuncUInt( const unsigned int& n )
 {
 	return n;
 }
 
-unsigned int 
+size_t
 hashFuncVoidPtr( void* const & pv )
 {
-   unsigned int ui = 0;
+   size_t ui = 0;
    for (int ix = 0; ix < (int)(sizeof(void*) / sizeof(int)); ++ix)
       {
       ui += ((unsigned int const*)&pv)[ix];
@@ -57,16 +57,16 @@ hashFuncVoidPtr( void* const & pv )
    return ui;
 }
 
-unsigned int 
+size_t
 hashFuncPROC_ID( const PROC_ID &procID )
 {
 	return ( (procID.cluster+(procID.proc*19)) );
 }
 
 // Chris Torek's world famous hashing function
-unsigned int hashFunction( char const *key )
+size_t hashFunction( char const *key )
 {
-    unsigned int i = 0;
+    size_t i = 0;
     if ( key ) {
 		for ( ; *key ; key++ ) {
 			i += (i<<5) + (const unsigned char)*key;
@@ -75,24 +75,24 @@ unsigned int hashFunction( char const *key )
     return i;
 }
 
-unsigned int hashFunction( const std::string &key )
+size_t hashFunction( const std::string &key )
 {
 	return hashFunction( key.c_str() );
 }
 
-unsigned int hashFunction( const MyString &key )
+size_t hashFunction( const MyString &key )
 {
 	return hashFunction( key.Value() );
 }
 
-unsigned int hashFunction( const YourString &key )
+size_t hashFunction( const YourString &key )
 {
 	return hashFunction( key.Value() );
 }
 
-unsigned int hashFunction( const YourStringNoCase &key )
+size_t hashFunction( const YourStringNoCase &key )
 {
-    unsigned int i = 0;
+    size_t i = 0;
 	for ( const char *p = key.Value(); *p ; p++ ) {
 		i += (i<<5) + (const unsigned char)(*p & ~0x20);
 	}
